Deduplicating plugins
